Review comments
- Made TaskConsumer interface, to avoid having cycles in the dependencies.
- Fixed javadoc error.
