+++ source/taskmanager/src/main/java/nl/aerius/taskmanager/domain/WorkerUpdateHandler.java
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see http://www.gnu.org/licenses/.
  */
-package nl.aerius.taskmanager;
+package nl.aerius.taskmanager.domain;
 
 /**
  * Handler to be informed when changed related to handling or tasks by workers and worker pool changes.
