+++ source/taskmanager/src/main/java/nl/aerius/taskmanager/scheduler/TaskScheduler.java
@@ -16,15 +16,14 @@
  */
 package nl.aerius.taskmanager.scheduler;
 
-import nl.aerius.taskmanager.WorkerUpdateHandler;
 import nl.aerius.taskmanager.domain.Task;
 import nl.aerius.taskmanager.domain.TaskQueue;
 import nl.aerius.taskmanager.domain.TaskSchedule;
+import nl.aerius.taskmanager.domain.WorkerUpdateHandler;
 
 /**
  * Interface for the scheduling algorithm. The implementation should maintain an internal list of all tasks added and return the task to be processed
- * in {@link #getNextTask()} based on whatever priority algorithm the scheduler implements. When a task is finished the scheduler receives
- * {@link WorkerUpdateHandler#onTaskFinished(Task)} event with the task finished.
+ * in {@link #getNextTask()} based on whatever priority algorithm the scheduler implements.
  */
 public interface TaskScheduler<T extends TaskQueue> extends WorkerUpdateHandler {
