+++ source/taskmanager/src/test/java/nl/aerius/taskmanager/TaskManagerTest.java
@@ -29,12 +29,11 @@ import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+import nl.aerius.taskmanager.MockTaskScheduler.MockSchedulerFactory;
 import nl.aerius.taskmanager.adaptor.AdaptorFactory;
 import nl.aerius.taskmanager.domain.PriorityTaskQueue;
 import nl.aerius.taskmanager.domain.PriorityTaskSchedule;
 import nl.aerius.taskmanager.domain.RabbitMQQueueType;
-import nl.aerius.taskmanager.scheduler.TaskScheduler.TaskSchedulerFactory;
-import nl.aerius.taskmanager.scheduler.fifo.FIFOTaskScheduler;
 import nl.aerius.taskmanager.scheduler.priorityqueue.PriorityTaskSchedulerFileHandler;
 
 /**
@@ -51,7 +50,7 @@ class TaskManagerTest {
   void setUp() throws IOException {
     executor = Executors.newCachedThreadPool();
     final AdaptorFactory factory = new MockAdaptorFactory();
-    final TaskSchedulerFactory<PriorityTaskQueue, PriorityTaskSchedule> schedulerFactory = new FIFOTaskScheduler.FIFOSchedulerFactory();
+    final MockSchedulerFactory schedulerFactory = new MockTaskScheduler.MockSchedulerFactory();
     taskManager = new TaskManager<>(executor, factory, schedulerFactory, factory.createWorkerSizeProvider());
     schedule = handler.read(new File(getClass().getClassLoader().getResource("queue/priority-task-scheduler.ops.json").getFile()));
   }
